requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
basicauth==0.4.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.4
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.32
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
Pillow==10.0.1
platformdirs==4.0.0
pre-commit==3.5.0
psycopg2-binary==2.9.8
pycryptodome==3.15.0
PyJWT==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.7
